Try some common periods first
Try some common periods before doing any other calculations. So now,
POI are not used for testing if a function is periodic if a common
period is found first.

Also, instead of trying to determine if a function is periodic by
testing if f(x)=f(x+period), we're testing if f(x+period)=f(x+2*period)
which is the same thing. But, this way it works for sec(x) and csc(x)
which for f(x+period) is translated to 1/cos(x) and 1/sin(x) and
equality wouldn't work.