For periodic functions, manually search within one period
When a function is determined as periodic and doing a manual search,
only search for POI within one period.

Also, if when searching manually for POI, those are too many (>10),
discard most of them and keep only the first, middle and last. Helps a
lot with functions with thousands of POI, like sin(1/cos(x)).