Don't use multiprocessing on win32
Since windows lacks os.fork(), creating multiple processes takes an
awful long time. Adding f(x)=x in linux takes a fraction of a second. On
windows, it takes about 16sec on the same PC. That sucks.

So, multiprocessing will be loaded conditionally on non-windows
platforms. On windows, threading is used instead. Now, because of the
Global Interpreter Lock in CPython, that means that multiple cores are
not actually utilized in windows. Once again, it sucks, but it's better
than taking a huge time to spawn multiple processes.